LwESP socket basics
This add support for sockets on top of the Netconn API. The code is based
on the LwIP implementation and adjusted for a Netconn variant provided by LwESP.

Currently only a basic functionality is provided including TCP socket
create, listen, bind, connect, read, write, and close. On the other hand
everything using UDP, select and poll will not work.